+++ ProjectAssignment_2/src/util/Rice.java
@@ -2,14 +2,14 @@ package util;
 
 public abstract class Rice extends Item{
     private double weight;
-    private int costInCents;
+    private double costInCents;
 
     @Override
     public double getCost(){
         return this.weight * this.costInCents;
     }
 
-    public Rice(String name, double weight, int costInCents){
+    public Rice(String name, double weight, double costInCents){
         super(name);
         this.weight = weight;
         this.costInCents = costInCents;
